requirements.txt
torch
torchvision
fastapi
uvicorn
python-multipart
opencv-python-headless
numpy
pyyaml
Pillow
jinja2
